% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcmatrix.R
\name{dcorgaussianbiv}
\alias{dcorgaussianbiv}
\title{Calculates distance correlation from Pearson correlation under assumption of a bivariate normal distribution}
\usage{
dcorgaussianbiv(rho)
}
\arguments{
\item{rho}{Pearson correlation.}
}
\value{
Distance correlation assuming a bivariate normal distribution
}
\description{
Calculates distance correlation from Pearson correlation under assumption of a bivariate normal distribution
}
