% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcsis.R
\name{dcsis}
\alias{dcsis}
\title{Performs Distance correlation Sure Independence Screening (DC-SIS)}
\usage{
dcsis(
  X,
  Y,
  k = floor(nrow(X)/log(nrow(X))),
  threshold = NULL,
  calc.cor = "spearman",
  calc.pval.cor = FALSE,
  return.data = FALSE,
  test = "none",
  adjustp = "none",
  b = 499,
  bias.corr = FALSE,
  use = "everything",
  algorithm = "auto"
)
}
\arguments{
\item{X}{A dataframe or matrix with n rows and p columns.}

\item{Y}{A vector-valued response of length n}

\item{k}{Number of variables that are selected (only used when threshold is not provided).}

\item{threshold}{If provided, variables with a distance correlation larger than threshold are selected.}

\item{calc.cor}{If set as "pearson", "spearman" or "kendall", a corresponding correlation matrix is addionally calculated.}

\item{calc.pval.cor}{IF TRUE, a p-value based on the Pearson or Spearman correlation matrix is calculated (not implemented for calc.cor ="kendall") using Hmisc::rcorr}

\item{return.data}{IF TRUE, X and Y are contained in the resulting dcmatrix object.}

\item{test}{specifies the type of test that is performed, "permutation" performs a Monte Carlo Permutation test. "gamma" performs a test based on a gamma approximation of the test statistic under the null. "conservative" performs a conservative two-moment approximation. "bb3" performs a quite precise three-moment approximation and is recommended when computation time is not an issue.}

\item{adjustp}{If setting this parameter to "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" or "fdr", corresponding adjusted p-values are returned.}

\item{b}{specifies the number of random permutations used for the permutation test. Ignored when test="gamma"}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance covariance should be calculated}

\item{use}{: "all" uses all observations, "complete.obs" excludes NA's, "pairwise.complete.obs" uses pairwise complete observations for each comparison.}

\item{affine}{logical; indicates if the affinely transformed distance covariance should be calculated or not.}

\item{algorithm:}{One of "auto", "fast", "memsave" and "standard". "memsave" is typically very inefficient for dcmatrix and should only be applied in exceptional cases.}
}
\value{
dcmatrix object
}
\description{
Performs Distance correlation Sure Independence Screening (DC-SIS)
}
