% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov_function.R
\name{distcov}
\alias{distcov}
\title{Calculate the distance covariance}
\usage{
distcov(
  X,
  Y,
  affine = FALSE,
  standardize = FALSE,
  bias.corr = FALSE,
  type.X = "sample",
  type.Y = "sample",
  metr.X = "euclidean",
  metr.Y = "euclidean",
  use = "all",
  algorithm = "auto"
)
}
\arguments{
\item{X}{contains either the first sample or its corresponding distance matrix. In the first case, this input can be either a vector of positive length, a matrix with one column or a data.frame with one column. In this case, type.X must be specified as "sample". In the second case, the input must be a distance matrix corresponding to the sample of interest. In this second case, type.X must be "distance".}

\item{Y}{see X.}

\item{affine}{logical; indicates if the affinely transformed distance covariance should be calculated or not.}

\item{standardize}{logical; should X and Y be standardized using the standard deviations of single observations?. No effect when affine = TRUE.}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance covariance should be calculated.}

\item{type.X}{For "distance", X is interpreted as a distance matrix. For "sample" (or any other value), X is intepreted as a sample}

\item{type.Y}{see type.X.}

\item{metr.X}{specifies the metric which should be used for X to analyse the distance covariance. TO DO: Provide details for this.}

\item{metr.Y}{see metr.X.}

\item{use}{: specifies how to treat missing values. "complete.obs" excludes NA's, "all" (or any other value) uses all observations.}

\item{algorithm}{: "fast" uses an O(n log n) algorithm if the observations are one-dimensional and metr.X and metr.Y are either "euclidean" or discrete. "memsave" uses a memory saving algorithm, which does not save the distance matrices. "standard" uses the classical algorithm. "auto" chooses the best algorithm for the specific setting using a rule of thumb.}
}
\value{
numeric giving the distance covariance between samples X and Y.
}
\description{
Calculate the distance covariance
}
