% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcovsurv.R
\name{ipcw.dcor}
\alias{ipcw.dcor}
\title{Calculates the inverse-probability-of-censoring weighted (IPCW) distance correlation}
\usage{
ipcw.dcor(Y, X, cutoff = NULL)
}
\arguments{
\item{Y}{A column with two rows, where the first row contains the survival times and the second row the status indicators (a survival object will work).}

\item{X}{A vector or matrix containing the covariate information.}

\item{cutoff}{If provided, all survival times larger than cutoff are set to the cutoff and all corresponding status indicators are set to one. Under most circumstances, choosing a cutoff is highly recommended.}
}
\value{
An inverse-probability of censoring weighted estimate for the distance correlation between X and the survival times.
}
\description{
Calculates the inverse-probability-of-censoring weighted (IPCW) distance correlation
}
