% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_function.R
\name{linked}
\alias{linked}
\title{Plots a linked chart for exploratory data analysis from a dcmatrix object using the functionality provided by the "rlc" package.}
\usage{
linked(
  dcmatrix,
  X = NULL,
  Y = NULL,
  heatmap = "dcor",
  heatmap0 = "dcor",
  size = 2,
  opacity = 1,
  discX = NULL,
  discY = NULL,
  jitt.disc = 0.1,
  jitt.cont = 0,
  smooths = "none",
  smooth.type = "loess",
  cgtabmax = NULL,
  subdat = NULL,
  subs = NULL,
  palette.heatmap = RColorBrewer::brewer.pal(9, "Reds"),
  palette.heatmap0 = RColorBrewer::brewer.pal(9, "YlGnBu"),
  coldom = NULL,
  coldom0 = NULL,
  ...
)
}
\arguments{
\item{dcmatrix}{A dcmatrix object}

\item{X}{should only be specified if the dcmatrix object does not contain the original data (see argument "return.data" of dcmatrix); should contain the same data as used in the calculation of the dcmatrix object.}

\item{Y}{see X.}

\item{heatmap}{specifies the dependence measure to use for the heatmap. Options are "dcor", "dcov", "logp" (for -log10 of the p-value), "cor" (for Pearson correlation) and "abscor" (for the absolute value of Pearson correlation).}

\item{heatmap0}{Only applicable if the dcmatrix object contains dependencies between objects that are not one-dimensional. In this case, two heatmaps are plotted, the first using the dependence meausure specified in heatmap0 and the dependencies between the groups in dcmatrix, the second using the dependence masure specified in heatmap and between the corresponding univariate variables.}

\item{size}{Passed to the options of the scatter plot.}

\item{opacity}{Passed to the options of the scatter plot.}

\item{discX}{Numeric vector specifying which columns in X should be interepreted as discrete variables. Factors and characters are always interpreted as discrete.}

\item{discY}{see discX}

\item{jitt.disc}{Jitter added to discrete variables.}

\item{jitt.cont}{Jitter added to continuous variables.}

\item{smooths}{specifies smooths that fitted to the scatter plots. Options include "none", "xtoy" (x as predictor, y as dependent variable), "ytox" (y as predictor, x as dependent variable) and "both".}

\item{smooth.type}{Type of the plotted smooths. Options are "loess", "spline" or "regline".}

\item{cgtabmax}{If set to a numeric value, a contingency table is shown for any pair of variables for which each of the variable has not more than cgtabmax unique values.}

\item{subdat}{optional; Data.frame with the same number of observations as the original data.}

\item{subs}{Names of some factor(!) variables in subdat. This allows to color the observations corresponding to these variables.}

\item{palette.heatmap}{Passed to the options of the heatmap.}

\item{palette.heatmap0}{Only applicable if the dcmatrix object contains dependencies between objects that are not one-dimensional. In this case, passed to the options of the corresponding heatmap.}

\item{...}{passed to the functions used for the smooths.}
}
\value{
A linked chart object opening in the browser. 

If the dcmatrix object contains only associations between univariate observations, this will consist of one heatmap with corresponding scatterplots.

If the dcmatrix contains associations between groups of variables, there are three windows. The first window displays a fixed heatmap contains the associations between the groups. The second window contains a changing heatmap containing the associations between single observations in selected groups. The third windows display a scatter plot between selected single observations.
}
\description{
Plots a linked chart for exploratory data analysis from a dcmatrix object using the functionality provided by the "rlc" package.
}
\examples{
A linked chart object opening in the browser.
X <-  matrix(rnorm(1000), ncol = 10)
dcm <- dcmatrix(X, return.data = TRUE)
linked(dcm)

X <-  matrix(rnorm(1000), ncol = 10)
dcm <- dcmatrix(X, return.data = TRUE, calc.dcor.pw = TRUE, group.X = c(rep("group1",3),rep("group2",3),rep("group3",4)))
linked(dcm)

Y <- matrix(rnorm(600), ncol = 6)
Y[,5] <- rbinom(100, 3, 0.5)
Y[,6] <- rbinom(100, 2, 0.3)
dcm <- dcmatrix(X, Y, return.data = TRUE, calc.dcor.pw = TRUE, group.X = c(rep("group1",3),rep("group2",3),rep("group3",4)), group.Y = c(rep("group1",4),rep("group2",1),rep("group3",1)), metr.Y = c("group1" = "euclidean", "group2" = "discrete", "group3" = "discrete"))
linked(dcm, discY = c(5,6))
}
