% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_dcmatrix.R
\name{plot.dcmatrix}
\alias{plot.dcmatrix}
\title{Plots a heatmap from a dcmatrix object using the function "pheatmap" from the package "pheatmap".}
\usage{
\method{plot}{dcmatrix}(
  x,
  type = "dcor",
  trunc.up = NULL,
  trunc.low = NULL,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  display_numbers = TRUE,
  ...
)
}
\arguments{
\item{x}{a dcmatrix object.}

\item{type}{specifies what should be displayed in the heatmap. One of "dcor", "dcov", "logp" (-log10 of corresponding p-values), "cor", "abscor" (absolute correlation), "logp.cor", "dcor.pw", "dcov.pw" or "logp.pw".}

\item{trunc.up}{truncates the values to be plotted; if set to numeric, all values larger than trunc.up are set to trunc.up.}

\item{trunc.low}{truncates the values to be plotted; if set to numeric, all values smaller than trunc.low are set to trunc.low.}

\item{cluster_rows, cluster_cols, display_numbers}{passed to pheatmap().}

\item{...}{passed to pheatmap().}
}
\value{
a heatmap plotting the entries of the slot specified in type of the object specified in dcmat.
}
\description{
Plots a heatmap from a dcmatrix object using the function "pheatmap" from the package "pheatmap".
}
