% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc_matrix2.R
\name{dcmatrix}
\alias{dcmatrix}
\title{Calculates distance covariance and distance correlation matrices}
\usage{
dcmatrix(
  X,
  Y = NULL,
  calc.dcov = TRUE,
  calc.dcor = TRUE,
  calc.cor = "none",
  calc.pval.cor = FALSE,
  return.data = TRUE,
  test = "none",
  adjustp = "none",
  b = 499,
  affine = FALSE,
  standardize = FALSE,
  bias.corr = FALSE,
  group.X = NULL,
  group.Y = NULL,
  metr.X = "euclidean",
  metr.Y = "euclidean",
  use = "everything",
  algorithm = "auto",
  fc.discrete = FALSE,
  calc.dcor.pw = FALSE,
  calc.dcov.pw = FALSE,
  test.pw = "none",
  metr.pw.X = "euclidean",
  metr.pw.Y = "euclidean"
)
}
\arguments{
\item{X}{A dataframe or matrix with n rows and p columns.}

\item{Y}{Either NULL or a dataframe or a matrix with n rows and q columns. If only X is provided, distance covariances/correlations are calculated between all groups in X. If X and Y are provided, distance covariances/correlations are calculated between all groups in X and all groups and Y.}

\item{calc.dcov}{Should distance covariance matrix be calculated?}

\item{calc.dcor}{Should distance correlation matrix be calculated?}

\item{calc.cor}{If set as "pearson", "spearman" or "kendall", a corresponding correlation matrix is addionally calculated.}

\item{calc.pval.cor}{IF TRUE, a p-value based on the Pearson or Spearman correlation matrix is calculated (not implemented for calc.cor ="kendall") using Hmisc::rcorr}

\item{return.data}{IF TRUE, X and Y are contained in the resulting dcmatrix object.}

\item{test}{specifies the type of test that is performed, "permutation" performs a Monte Carlo Permutation test. "gamma" performs a test based on a gamma approximation of the test statistic under the null. "conservative" performs a conservative two-moment approximation. "bb3" performs a quite precise three-moment approximation and is recommended when computation time is not an issue.}

\item{adjustp}{If setting this parameter to "holm", "hochberg", "hommel", "bonferroni", "BH", "BY" or "fdr", corresponding adjusted p-values are returned.}

\item{b}{specifies the number of random permutations used for the permutation test. Ignored when test="gamma"}

\item{affine}{logical; indicates if the affinely transformed distance covariance should be calculated or not.}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance covariance should be calculated}

\item{group.X}{A vector of length p, each entry specifying the group membership of the respective column in X. Each group is handled as one sample for calculating the distance covariance/correlation matrices. If NULL, every sample is handled as an individual group.}

\item{group.Y}{A vector of length q, each entry specifying the group membership of the respective column in Y. Each group is handled as one sample for calculating the distance covariance/correlation matrices. If NULL, every sample is handled as an individual group.}

\item{metr.X}{Either a single metric or a list providing a metric for each group in X.}

\item{metr.Y}{see metr.X.}

\item{use}{: "all" uses all observations, "complete.obs" excludes NA's, "pairwise.complete.obs" uses pairwise complete observations for each comparison.}

\item{algorithm:}{One of "auto", "fast", "memsave" and "standard". "memsave" is typically very inefficient for dcmatrix and should only be applied in exceptional cases.}

\item{fc.discrete:}{If TRUE, "discrete" metric is applied automatically on samples corresponding to a single column of type "factor" or "character".}
}
\value{
dcmatrix object
}
\description{
Calculates distance covariance and distance correlation matrices
}
