% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcov_test2.R
\name{distcov.test}
\alias{distcov.test}
\title{Performs a distance covariance test}
\usage{
distcov.test(
  X,
  Y,
  method = "permutation",
  b = 499L,
  ln = 20,
  affine = FALSE,
  standardize = FALSE,
  bias.corr = FALSE,
  type.X = "sample",
  type.Y = "sample",
  metr.X = "euclidean",
  metr.Y = "euclidean",
  use = "all",
  return.data = FALSE,
  algorithm = "auto"
)
}
\arguments{
\item{X}{contains either the first sample or its corresponding distance matrix. In the first case, this input can be either a vector of positive length, a matrix with one column or a data.frame with one column. In this case, type.X must be specified as "sample". In the second case, the input must be a distance matrix corresponding to the sample of interest. In this second case, type.X must be "distance".}

\item{Y}{see X.}

\item{b}{specifies the number of random permutations/bootstrap samples used for the permutation or wild bootstraps tests. Ignored for other tests.}

\item{ln}{block size parameter for wild bootstrap tests. Ignored for other tests.}

\item{affine}{logical; indicates if the affinely transformed distance covariance should be calculated or not.}

\item{bias.corr}{logical; indicates if the bias corrected version of the sample distance covariance should be calculated.}

\item{type.X}{either "sample" or "distance"; specifies the type of input for X.}

\item{type.Y}{see type.X.}

\item{metr.X}{specifies the metric which should be used for X to analyse the distance covariance. One of "euclidean", "gaussian", "boundsq", "alpha", "discrete" or a user-defined metric or kernel (see examples). It is also possible to specify parameters for "gaussian", "boundsq" and "alpha" via c(metr,param), e.g. c("gaussian",2) will result in a Gaussian metric with bandwith parameter 2.}

\item{metr.Y}{see metr.X.}

\item{use}{: "all" uses all observations, "complete.obs" excludes NA's}

\item{return.data}{: IF TRUE, X and Y are contained in the resulting distcov.test object.}

\item{test}{specifies the type of test that is performed, "permutation" performs a Monte Carlo Permutation test. "gamma" performs a test based on a gamma approximation of the test statistic under the null. "conservative" performs a conservative two-moment approximation. "bb3" performs a quite precise three-moment approximation and is recommended when computation time is not an issue.}

\item{algorithm:}{One of "auto", "fast", "memsave" and "standard". "fast" has a lower computational complexity than "memsave" and "standard", but cannot be used with vector-valued data, other metrics than "discrete" or "euclidean" or the "bb3" test. "memsave" can be used in most settings and is slightly faster than the "standard" algorithm. Moreover, different to "standard", it can be used for very large sample size. "standard" can be used with all metrics and tests, however is the slowest algorithm and breaks down with large sample sizes. "auto" chooses an algorithm using feasibility and rules of thumb.}
}
\value{
distcov.test object
}
\description{
Performs a distance covariance test
}
