% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distcovsurv.R
\name{ipcw.dcov.test}
\alias{ipcw.dcov.test}
\title{Performs a permutation test based on the IPCW distance covariance}
\usage{
ipcw.dcov.test(Y, X, cutoff = NULL, B = 499)
}
\arguments{
\item{Y}{A column with two rows, where the first row contains the survival times and the second row the status indicators (a survival object will work).}

\item{X}{A vector or matrix containing the covariate information.}

\item{cutoff}{If provided, all survival times larger than cutoff are set to the cutoff and all corresponding status indicators are set to one. Under most circumstances, choosing a cutoff is highly recommended.}

\item{B}{The number of permutations used for the permutation test}
}
\value{
An list with two arguments, $dcov contains the IPCW distance covariance, $pvalue the corresponding p-value
}
\description{
Performs a permutation test based on the IPCW distance covariance
}
